+++ demo/scripts/use_metadata_tools.py
@@ -9,17 +9,17 @@
 #
 #################################################################
 
-from czitools.czi_metadata import CziMetadata, writexml, get_metadata_as_object, obj2dict
-from czitools.metadata.dimension import CziDimensions
-from czitools.metadata.boundingbox import CziBoundingBox
-from czitools.metadata.channel import CziChannelInfo
-from czitools.metadata.scaling import CziScaling
-from czitools.metadata.sample import CziSampleInfo
-from czitools.metadata.objective import CziObjectives
-from czitools.metadata.microscope import CziMicroscope
-from czitools.metadata.add_metadata import CziAddMetaData
-from czitools.metadata.detector import CziDetector
-from czitools.tools import misc
+from czitools.metadata_tools.czi_metadata import CziMetadata, writexml, get_metadata_as_object, obj2dict
+from czitools.metadata_tools.dimension import CziDimensions
+from czitools.metadata_tools.boundingbox import CziBoundingBox
+from czitools.metadata_tools.channel import CziChannelInfo
+from czitools.metadata_tools.scaling import CziScaling
+from czitools.metadata_tools.sample import CziSampleInfo
+from czitools.metadata_tools.objective import CziObjectives
+from czitools.metadata_tools.microscope import CziMicroscope
+from czitools.metadata_tools.add_metadata import CziAddMetaData
+from czitools.metadata_tools.detector import CziDetector
+from czitools.utils import misc
 from pathlib import Path
 
 # adapt to your needs
@@ -27,15 +27,15 @@ defaultdir = Path(Path(__file__).resolve().parents[2]) / "data"
 
 # open s simple dialog to select a CZI file
 filepath = misc.openfile(directory=str(defaultdir),
-                               title="Open CZI Image File",
-                               ftypename="CZI Files",
-                               extension="*.czi")
+                         title="Open CZI Image File",
+                         ftypename="CZI Files",
+                         extension="*.czi")
 print(filepath)
 
-# get the metadata at once as one big class
+# get the metadata_tools at once as one big class
 mdata = CziMetadata(filepath)
 
-# get only specific metadata
+# get only specific metadata_tools
 czi_dimensions = CziDimensions(filepath)
 print("SizeS: ", czi_dimensions.SizeS)
 print("SizeT: ", czi_dimensions.SizeT)
@@ -50,7 +50,7 @@ xmlfile = writexml(filepath)
 # get info about the channels
 czi_channels = CziChannelInfo(filepath)
 
-# get the complete metadata from the CZI as one big object
+# get the complete metadata_tools from the CZI as one big object
 czimd_complete = get_metadata_as_object(filepath)
 
 # get an object containing only the dimension information
@@ -83,7 +83,7 @@ print(czi_bbox.scenes_bounding_rect)
 print(czi_bbox.total_rect)
 print(czi_bbox.total_bounding_box["C"])
 
-# get selected metadata as a dictionary
+# get selected metadata_tools as a dictionary
 mdata_dict = obj2dict(mdata)
 
 # and convert to pd.DataFrame
