+++ demo/scripts/use_read_tools.py
@@ -11,7 +11,7 @@
 
 from czitools import read_tools
 from czitools import napari_tools
-from czitools.tools import misc
+from czitools.utils import misc
 import napari
 from pathlib import Path
