+++ src/czitools/_tests/test_stageXY.py
@@ -1,4 +1,4 @@
-from czitools import czi_metadata as czimd
+from czitools.metadata_tools import czi_metadata as czimd
 from pathlib import Path
 import pytest
 from typing import Dict
@@ -28,7 +28,7 @@ def test_stage_xy(czifile: str, results: Dict) -> None:
     # get the CZI filepath
     filepath = basedir / "data" / czifile
 
-    # read the metadata
+    # read the metadata_tools
     md = czimd.CziMetadata(filepath)
 
     if md.image.SizeS is not None:
