+++ src/czitools/metadata_tools/attachment.py
@@ -2,7 +2,7 @@ from typing import Optional, Union, List
 from dataclasses import dataclass, field
 from box import Box
 import os
-from czitools.tools.logger import get_logger
+from czitools.utils.logger import get_logger
 from pathlib import Path
 import validators
