+++ src/czitools/metadata_tools/helper.py
