+++ src/czitools/metadata_tools/scaling.py
@@ -3,9 +3,9 @@ from dataclasses import dataclass, field
 from box import Box, BoxList
 import os
 import numpy as np
-from czitools.tools.logger import get_logger
-from czitools.tools.box import get_czimd_box
-from czitools.metadata.helper import ValueRange
+from czitools.utils.logger import get_logger
+from czitools.utils.box import get_czimd_box
+from czitools.metadata_tools.helper import ValueRange
 
 logger = get_logger()
