+++ src/czitools/napari_tools/napari_tools.py
@@ -10,8 +10,8 @@
 #################################################################
 
 import sys
-from czitools.tools import logger
-from czitools.metadata.helper import ValueRange
+from czitools.utils import logger
+from czitools.metadata_tools.helper import ValueRange
 
 logger = logger.get_logger()
 
@@ -34,9 +34,9 @@ else:
     from PyQt5.QtCore import Qt
     from PyQt5 import QtWidgets
     from PyQt5.QtGui import QFont
-    from czitools import czi_metadata as czimd
-    from czitools.tools import misc
-    from czitools.tools.datatreewiget import DataTreeWidget
+    from czitools.metadata_tools import czi_metadata as czimd
+    from czitools.utils import misc
+    from czitools.utils.datatreewiget import DataTreeWidget
     import numpy as np
     from typing import (
         List,
@@ -69,13 +69,13 @@ class MdTableWidget(QWidget):
         header.setDefaultAlignment(Qt.AlignLeft)
 
     def update_metadata(self, md_dict: Dict) -> None:
-        """Update the table with the metadata from the dictionary
+        """Update the table with the metadata_tools from the dictionary
 
         Args:
             md_dict (Dict): Metadata dictionary
         """
 
-        # number of rows is set to number of metadata entries
+        # number of rows is set to number of metadata_tools entries
         row_count = len(md_dict)
         col_count = 2
         self.mdtable.setColumnCount(col_count)
@@ -83,7 +83,7 @@ class MdTableWidget(QWidget):
 
         row = 0
 
-        # update the table with the entries from metadata dictionary
+        # update the table with the entries from metadata_tools dictionary
         for key, value in md_dict.items():
             newkey = QTableWidgetItem(key)
             self.mdtable.setItem(row, 0, newkey)
@@ -141,7 +141,7 @@ def show(
     dask_cache_size: Annotated[float, ValueRange(0.5, 0.9)] = 0.5,
 ) -> List:
     """Display the multidimensional array inside the Napari viewer.
-    Optionally the CziMetadata class will be used show a table with the metadata.
+    Optionally the CziMetadata class will be used show a table with the metadata_tools.
     Every channel will be added as a new layer to the viewer.
 
     Args:
@@ -153,9 +153,9 @@ def show(
         contrast (Literal[str], optional): method to be used to calculate an appropriate display scaling.
             - "calc" : real min & max calculation (might be slow) be calculated (slow)
             - "napari_auto" : Let Napari figure out a display scaling. Will look in the center of an image!
-            - "from_czi" : use the display scaling from ZEN stored inside the CZI metadata. Defaults to "calc".
+            - "from_czi" : use the display scaling from ZEN stored inside the CZI metadata_tools. Defaults to "calc".
         gamma (float, optional): gamma value for the Viewer for all layers Defaults to 0.85.
-        show_metadata (mdviewoption, optional): Option to show metadata as tree or table. Defaults to "tree".
+        show_metadata (mdviewoption, optional): Option to show metadata_tools as tree or table. Defaults to "tree".
         name_sliders (bool, optional): option to use the dimension letters as slider labels for the viewer. Defaults to False.
         dask_cache_size(float, optional): option to resize the dask cache used for opportunistic caching. Range [0 - 1]
 
@@ -191,13 +191,13 @@ def show(
     scalefactors[dim_order["Z"]] = metadata.scale.ratio["zx_sf"] * 1.001
 
     if show_metadata.lower != "none":
-        # add PyQTGraph DataTreeWidget to Napari viewer to show the metadata
+        # add PyQTGraph DataTreeWidget to Napari viewer to show the metadata_tools
         if show_metadata == "tree":
             md_dict = czimd.create_md_dict_nested(metadata, sort=True, remove_none=True)
             mdtree = MdTreeWidget(data=md_dict, expandlevel=1)
             viewer.window.add_dock_widget(mdtree, name="MetadataTree", area="right")
 
-        # add QTableWidget DataTreeWidget to Napari viewer to show the metadata
+        # add QTableWidget DataTreeWidget to Napari viewer to show the metadata_tools
         if show_metadata == "table":
             md_dict = czimd.create_md_dict_red(metadata, sort=True, remove_none=True)
             mdtable = MdTableWidget()
@@ -216,7 +216,7 @@ def show(
         try:
             # get the channel name
             chname = metadata.channelinfo.names[ch]
-            # inside the CZI metadata colors are defined as ARGB hexstring
+            # inside the CZI metadata_tools colors are defined as ARGB hexstring
             rgb = "#" + metadata.channelinfo.colors[ch][3:]
             ncmap = Colormap(["#000000", rgb], name="cm_" + chname)
         except (KeyError, IndexError) as e:
@@ -321,7 +321,7 @@ def show(
 
 
 def rename_sliders(sliders: Tuple, dim_order: Dict) -> Tuple:
-    """Rename the sliders inside the Napari viewer based on the metadata
+    """Rename the sliders inside the Napari viewer based on the metadata_tools
 
 
     Args:
