+++ src/czitools/utils/datatreewiget.py
